moved functionality into class
